# v2.0.0: Complete Rust Rewrite with 10-30x Performance Improvements