--- a/.gitignore
+++ b/.gitignore
@@ -42,3 +42,5 @@ target/
 Cargo.lock
 **/*.rs.bk
 *.pdb
+*.so
+*.dylib