--- a/crates/html-to-markdown-py/python/html_to_markdown/__init__.py
+++ b/crates/html-to-markdown-py/python/html_to_markdown/__init__.py
@@ -1,0 +1,5 @@
+"""Python bindings for html-to-markdown Rust library."""
+
+from html_to_markdown._html_to_markdown import convert
+
+__all__ = ["convert"]