--- a/crates/html-to-markdown/src/lib.rs
+++ b/crates/html-to-markdown/src/lib.rs
@@ -1,0 +1,8 @@
+//! # html-to-markdown
+//!
+//! A modern, type-safe library for converting HTML to Markdown with comprehensive HTML5 support.
+
+pub fn convert(html: &str) -> String {
+    // TODO: Implementation
+    html.to_string()
+}