… .github/actions/build-wheels/action.yaml
… .github/workflows/ci.yaml
… .github/workflows/test-wheels.yaml
… .pre-commit-config.yaml
… BASELINE_BENCHMARKS.md
… CHANGELOG.md
… Cargo.toml
… IMPLEMENTATION_PLAN.md
… README.md
… RUST_CODE_REVIEW.md
… RUST_DEPENDENCIES.md
… Taskfile.yaml
… V2_API_DESIGN.md
… V2_PERFORMANCE_RESULTS.md
… WHITESPACE_ANALYSIS.md
… crates/html-to-markdown-cli/Cargo.toml
… crates/html-to-markdown-py/Cargo.toml
… crates/html-to-markdown-py/pyproject.toml
… crates/html-to-markdown-py/python/html_to_markdown/_html_to_markdown.abi3.so
… crates/html-to-markdown-py/python/html_to_markdown/_html_to_markdown.pyi
… crates/html-to-markdown-py/uv.lock
… crates/html-to-markdown/Cargo.toml
… crates/html-to-markdown/examples/basic.rs
… crates/html-to-markdown/examples/table.rs
… crates/html-to-markdown/examples/test_escape.rs
… crates/html-to-markdown/examples/test_inline_formatting.rs
… crates/html-to-markdown/examples/test_lists.rs
… crates/html-to-markdown/examples/test_semantic_tags.rs
… crates/html-to-markdown/examples/test_tables.rs
… crates/html-to-markdown/examples/test_task_lists.rs